day3 pt2

+++ lib/Day3/Day3.h
@@ -4,7 +4,7 @@
 #include <Arduino.h>
 
 int solve_day3_part1();
-int solve_day3_part2();
+long long solve_day3_part2();
 
 void solve_day3();
